requirements.txt
Django~=4.0.5
python-decouple
bs4
requests
django-bootstrap-v5
django-crispy-forms
crispy-bootstrap5
pandas
